mtd-utils: flash_speed: Try to measure RWW earned latency
Reads While Writes (RWW) allow reads to be performed on a different bank
while an internal write or erase is happening.

The main benefit is the possible reduced latency when reading a page
while writing another one.

We can try to compare the positive impact of such a feature by enhancing
the flash_speed test tool with the following test:
- Measure the time taken by an eraseblock write followed by an
  eraseblock read.
- Measure the time taken by an eraseblock write in parallel with an
  eraseblock read.
- The difference between the two is the latency saved with the RWW
  feature.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>